refactoring - uniqueness validator plugin - validations class for a given model - validate email & slack_id
