+++ models/index.ts
@@ -2,4 +2,4 @@
  * Import any models that should be instantiated here.
  * Their schema's will be built the first time
  */
-import './User.model';
+import './User/User.model';
